refactor: rename files
simpler and easier to read. can tell it is a router just because it's in routes folder
