+++ src/routes/posts.ts
@@ -10,7 +10,7 @@ import {
 } from "@/controllers/posts";
 import { attachPost } from "@/middleware/attach";
 import { authenticateWithJwt, checkIsAuthor } from "../middleware/auth";
-import commentsRouter from "./commentsRouter";
+import commentsRouter from "./comments";
 
 const postsRouter = Router();
